+++ source/port/serial-port-posix.c
@@ -24,13 +24,13 @@ void* serial_port_open(const char *name)
 
 int serial_port_close(void* port)
 {
-	int fd = (intptr_t)port;
+	int fd = (int)(intptr_t)port;
 	return close(fd);
 }
 
 int serial_port_flush(void* port)
 {
-	int fd = (intptr_t)port;
+	int fd = (int)(intptr_t)port;
 	return tcflush(fd, TCIOFLUSH);
 }
 
@@ -65,7 +65,7 @@ static int serial_port_baud_rate(int baudrate)
 
 int serial_port_setattr(void* port, int baudrate, int databits, int parity, int stopbits, int flowctrl)
 {
-	int fd = (intptr_t)port;
+	int fd = (int)(intptr_t)port;
 	struct termios options;
 
 	baudrate = serial_port_baud_rate(baudrate);
@@ -164,12 +164,12 @@ int serial_port_setattr(void* port, int baudrate, int databits, int parity, int 
 
 int serial_port_write(void* port, const void* data, int bytes)
 {
-	int fd = (intptr_t)port;
-	return write(fd, data, bytes);
+	int fd = (int)(intptr_t)port;
+	return (int)write(fd, data, bytes);
 }
 
 int serial_port_read(void* port, void* data, int bytes)
 {
-	int fd = (intptr_t)port;
-	return read(fd, data, bytes);
+	int fd = (int)(intptr_t)port;
+	return (int)read(fd, data, bytes);
 }
