+++ source/uuid.c
@@ -48,7 +48,7 @@ static int read_random_bytes(void* buf, int len)
 
 	for (r = n = 0; n < len; n += r)
 	{
-		r = fread((uint8_t*)buf + n, 1, len - n, fp);
+		r = (int)fread((uint8_t*)buf + n, 1, len - n, fp);
 		if (r <= 0)
 			break;
 	}
